requirements.txt
fastapi
uvicorn
jinja2
httpx
PyJWT
